feat(components): add NewActivity component to show/hide activity card
This commit introduces the new `NewActivity` component.

The component renders a clickable `<h2>` heading that acts as a button to initiate the creation of a new activity. It uses the `useState` hook to manage the visibility of a `Card` component. Clicking the heading toggles the display of this card, which will be used for the activity creation form.